requirements.txt
# requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
jinja2==3.1.6
python-dotenv==1.1.1
sentence-transformers==5.0.0
faiss-cpu==1.11.0
PyMuPDF==1.26.3
python-docx==1.2.0
Pillow==11.3.0
pytesseract==0.3.13
requests==2.32.4
numpy==2.3.1
beautifulsoup4==4.13.4
trafilatura==2.0.0
lxml==6.0.0